expense-tracker2/backend/models.py
from pymongo import MongoClient
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

# MongoDB connection
MONGO_URI = os.getenv('MONGO_URI')
client = MongoClient(MONGO_URI)
db = client['expense_tracker']

# Collections
users_collection = db['users']
expenses_collection = db['expenses']