expense-tracker2/backend/requirements.txt
Flask==2.3.2
pymongo==4.4.0
flask-cors==4.0.0
python-dotenv==1.0.0